#32 Refactor ArticleService with ArticleQueryService
