+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/domain/ArticleService.kt
@@ -1,7 +1,5 @@
 package io.github.raeperd.realworldspringbootkotlin.domain
 
-import org.springframework.data.domain.Page
-import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.Instant
@@ -47,14 +45,6 @@ class ArticleService(
         }
         articleRepository.deleteArticle(article)
     }
-
-    fun getArticles(pageable: Pageable, param: ArticleQueryParam, viewerId: Long? = null): Page<ArticleDTO> {
-        val articles = articleRepository.getAllArticles(pageable, param)
-        return viewerId?.let { id -> userRepository.findUserByIdOrThrow(id) }
-            ?.let { user -> articles.map { it.toArticleDTO(user) } }
-            ?: articles.map { it.toArticleDTO() }
-    }
-
 }
 
 data class ArticleDTO(
