+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/web/ArticleRestController.kt
@@ -9,7 +9,8 @@ import org.springframework.web.bind.annotation.*
 
 @RestController
 class ArticleRestController(
-    private val articleService: ArticleService
+    private val articleService: ArticleService,
+    private val articleQueryService: ArticleQueryService
 ) {
     @ResponseStatus(CREATED)
     @PostMapping("/articles")
@@ -39,7 +40,7 @@ class ArticleRestController(
 
     @GetMapping("/articles")
     fun getArticles(pageable: Pageable, param: ArticleQueryParam, payload: JWTPayload?): MultipleArticleModel {
-        return articleService.getArticles(pageable, param)
+        return articleQueryService.getArticles(pageable, param)
             .toMultipleArticleModel()
     }
